refac(zk): move up `CommitXXX` from `ProverBase` to `Entity`
In addition, this also sorts the methods in the order of poly, evals
and containers.
