+++ tachyon/zk/base/entities/BUILD.bazel
@@ -7,6 +7,7 @@ tachyon_cc_library(
     hdrs = ["entity.h"],
     deps = [
         "//tachyon/base:logging",
+        "//tachyon/crypto/commitments:vector_commitment_scheme_traits_forward",
         "//tachyon/crypto/transcripts:transcript",
         "//tachyon/zk/base:row_index",
     ],
@@ -18,7 +19,6 @@ tachyon_cc_library(
     deps = [
         ":entity",
         "//tachyon/base:logging",
-        "//tachyon/crypto/commitments:vector_commitment_scheme_traits_forward",
         "//tachyon/zk/base:blinded_polynomial",
         "//tachyon/zk/base:blinder",
         "//tachyon/zk/base:row_index",
