+++ tachyon/zk/plonk/halo2/verifier.h
@@ -170,9 +170,7 @@ class Verifier : public VerifierBase<PCS> {
     return base::Map(columns, [this](const Evals& column) {
       std::vector<F> expanded_evals = column.evaluations();
       expanded_evals.resize(this->pcs_.N());
-      Commitment c;
-      CHECK(this->pcs_.CommitLagrange(Evals(std::move(expanded_evals)), &c));
-      return c;
+      return this->Commit(expanded_evals);
     });
   }
